+++ BACKEND/src/server.js
@@ -6,6 +6,7 @@ const { iniciarJobColeta } = require('./jobs/coletaJob');
 const coletaMassivasRoutes = require('./routes/coletaMassivasRoutes');
 const { iniciarJobMassivas } = require('./jobs/coletaMassivasJob');
 const { iniciarJobScalefusion } = require('./jobs/scalefusionJob');
+const { iniciarJobSegsatFrota } = require('./jobs/segsatFrotaJob');
 const authRoutes = require('./routes/authRoutes');
 const usuariosRoutes = require('./routes/usuariosRoutes');
 const dashboardRoutes = require('./routes/dashboardRoutes');
@@ -52,4 +53,5 @@ app.listen(PORT, () => {
   iniciarJobColeta();
   iniciarJobMassivas();
   iniciarJobScalefusion();
+  iniciarJobSegsatFrota();
 });